added .gitignore rule
