+++ .gitignore
@@ -161,3 +161,4 @@ cython_debug/
 
 # Private info
 credentials.json
+test-credentials.json
