--- test-credentials.json
@@ -1,1 +1,0 @@
-{ "token": "" }
