+++ test/helper.py
@@ -1,0 +1,2 @@
+test_bot_id =     "1226725123378118726"
+test_channel_id = "1227809613198528542"
